+++ sdk/src/main/java/com/walmartlabs/concord/sdk/Constants.java
@@ -135,6 +135,11 @@ public class Constants {
          * Enables additional debug logging for various states of execution.
          */
         public static final String DEBUG_KEY = "debug";
+
+        /**
+         * Schedules a process on the specified date and time.
+         */
+        public static final String START_AT_KEY = "startAt";
     }
 
     /**